+++ model/gorm/paper.go
@@ -2,7 +2,7 @@ package model
 
 // Paper : table `paper`
 type Paper struct {
-	ID             int    `gorm:"type : int(20); primary_key; not null; auto_increment"`
+	ID             int64  `gorm:"type : int(20); primary_key; not null; auto_increment"`
 	Difficulty     int    `gorm:"type : tinyint(4); not null"`
 	KnowledgePoint string `gorm:"type : text; not null"`
 }
