+++ model/gorm/user_class.go
@@ -2,7 +2,7 @@ package model
 
 // UserClass : teble `user`
 type UserClass struct {
-	UserID       int    `gorm:"type : int(20); not null"`
+	UserID       int64  `gorm:"type : int(20); not null"`
 	ClassID      int    `gorm:"type : int(20); not null"`
 	announcement []byte `gorm:"type : blob; default null"`
 }
